remove FBX support and update dependencies.zip!
the autodesk FBX SDK used by the game was clunky and old, and FBX isn't even used in the base game - if this is reimplemented in the future it would be best to use a more modern solution but I don't see any urgency to do so. this should make upgrading to a more modern version of Visual Studio easier!